refactoring  test, product service and globalHAndleException

+++ domain/ordermgmt-product/src/main/java/com/ordermgmt/products/service/ProductService.java
@@ -39,6 +39,9 @@ public class ProductService implements IProductService{
 
     @Override
     public void deleteProduct(Long id) {
+        if (!iProductRepository.existsById(id)) {
+            throw new ProductNotFoundException("Product not found with id: " + id);
+        }
         iProductRepository.deleteById(id);
 
     }
